+++ src/main/java/org/salt/jlangchain/core/agent/AgentExecutor.java
@@ -21,6 +21,7 @@ import org.salt.function.flow.Info;
 import org.salt.function.flow.context.ContextBus;
 import org.salt.jlangchain.core.BaseRunnable;
 import org.salt.jlangchain.core.ChainActor;
+import org.salt.jlangchain.core.agent.memory.*;
 import org.salt.jlangchain.core.common.CallInfo;
 import org.salt.jlangchain.core.handler.TranslateHandler;
 import org.salt.jlangchain.core.history.memory.ConversationMemoryStorerBase;
