+++ src/main/java/org/salt/jlangchain/core/agent/memory/AgentContext.java
@@ -12,7 +12,10 @@
  * limitations under the License.
  */
 
-package org.salt.jlangchain.core.agent;
+package org.salt.jlangchain.core.agent.memory;
+
+import org.salt.jlangchain.core.agent.AgentExecutor;
+import org.salt.jlangchain.core.agent.McpAgentExecutor;
 
 /**
  * Factory for per-invocation {@link AgentTaskContext} instances.
