+++ src/main/java/org/salt/jlangchain/core/agent/memory/AgentTaskContext.java
@@ -12,7 +12,7 @@
  * limitations under the License.
  */
 
-package org.salt.jlangchain.core.agent;
+package org.salt.jlangchain.core.agent.memory;
 
 import org.salt.jlangchain.core.message.BaseMessage;
 import org.salt.jlangchain.core.prompt.value.ChatPromptValue;
