+++ src/test/java/org/salt/jlangchain/demo/flow/ChainAgentContextTest.java
@@ -21,10 +21,10 @@ import org.salt.jlangchain.TestApplication;
 import org.salt.jlangchain.core.ChainActor;
 import org.salt.jlangchain.core.agent.AgentExecutor;
 import org.salt.jlangchain.core.agent.McpAgentExecutor;
-import org.salt.jlangchain.core.agent.SlidingWindowContext;
+import org.salt.jlangchain.core.agent.memory.SlidingWindowContext;
 import org.salt.jlangchain.core.history.HistoryInfos;
 import org.salt.jlangchain.core.history.memory.buffer.ConversationBufferMemoryStorer;
-import org.salt.jlangchain.core.history.storage.InMemoryAgentTaskStorage;
+import org.salt.jlangchain.core.agent.storage.InMemoryAgentTaskStorage;
 import org.salt.jlangchain.core.history.storage.InMemoryConversationStorage;
 import org.salt.jlangchain.core.llm.aliyun.ChatAliyun;
 import org.salt.jlangchain.core.parser.generation.ChatGeneration;
